Staff Feature

+++ app/Http/Controllers/Admin/BrandController.php
@@ -75,6 +75,7 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
+        abort_if(request()->user()->role_id, 403, 'Not Allowed.');
         $brand->delete();
         return redirect()
             ->action([self::class, 'index'])
